uses bit.ly to make github parser happy, which prevents a deployment
The page build failed for the `gh-pages` branch with the following error:

The tag `{%` on line 83 in `blog/2018-06-apollo-graphql-code-generator/README.md` was not properly closed with `%}`. For more information, see https://help.github.com/articles/page-build-failed-tag-not-properly-terminated/.

For information on troubleshooting Jekyll see:

  https://help.github.com/articles/troubleshooting-jekyll-builds

If you have any questions you can contact us by replying to this email.